refactor: pointer-safe runtime contracts; remove the network code (v0.3.0 WP1)
Pointer-width fixes that do not change the save format (that is WP2):

- RS2OpaqueData (uintptr_t, RS2Width.h) replaces the DWORD slots that
  carried object pointers: CListElement data, CDragContainer data and
  the CMenuCommander Dispatch / DoubleClick data, with every caller.
  One slot holds several pointer types and, in one place, an option
  number, so it stays a generic slot rather than a typed pointer.
- _findfirst handles are intptr_t (four enumeration paths), compared
  against -1.
- The DirectInput axis-enumeration context carries the joystick number
  through INT_PTR explicitly.
- The VC6-era "typedef unsigned long *DWORD_PTR" is gone; the SDK's
  pointer-sized types are asserted instead.
- SelectFile clears the buffer by its length, not by sizeof(pointer).
- size_t / pointer-difference narrowing goes through checked helpers
  (RS2SizeToInt, RS2SizeToDword, RS2DiffToInt: logged and clamped when
  out of range), including two fprintf("%d", size()) calls that passed
  a 64-bit value; LoadBinaryText refuses an unusable ftell result.

Network.cpp / Network.h are deleted (user decision): the ~20 callers
that still consult the network state link against RS2NetworkDisabled.cpp,
where no session can exist.

x64 width warnings: 155 -> 3 (the save-file address map, WP2).

+++ md5.cpp
@@ -1,3 +1,4 @@
+//	Modified for RS2EX on 2026-09-26.
 // MD5.CC - source code for the C++/object oriented translation and 
 //          modification of MD5.
 
@@ -122,7 +123,7 @@ void MD5::update(FILE *file){
   unsigned char buffer[1024];
   int len;
 
-  while (len=fread(buffer, 1, 1024, file))
+  while (len=(unsigned int)fread(buffer, 1, 1024, file))	/* [RS2EX] at most 1024 */
     update(buffer, len);
 
   fclose (file);
